requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
beautifulsoup4==4.14.3
    # via -r requirements.in
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.7
    # via requests
click==8.3.1
    # via uvicorn
distro==1.9.0
    # via openai
fastapi==0.135.3
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   -r requirements.in
    #   langgraph-sdk
    #   langsmith
    #   openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.13.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.1.1
    # via jsonpatch
langchain==1.2.15
    # via -r requirements.in
langchain-core==1.2.25
    # via
    #   langchain
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langgraph==1.1.5
    # via
    #   -r requirements.in
    #   langchain
langgraph-checkpoint==4.0.1
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.9
    # via langgraph
langgraph-sdk==0.3.12
    # via langgraph
langsmith==0.7.25
    # via langchain-core
openai==2.30.0
    # via -r requirements.in
orjson==3.11.8
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.12.2
    # via langgraph-checkpoint
packaging==26.0
    # via
    #   langchain-core
    #   langsmith
pydantic==2.12.5
    # via
    #   -r requirements.in
    #   fastapi
    #   langchain
    #   langchain-core
    #   langgraph
    #   langsmith
    #   openai
pydantic-core==2.41.5
    # via pydantic
python-dotenv==1.2.2
    # via
    #   -r requirements.in
    #   uvicorn
pyyaml==6.0.3
    # via
    #   langchain-core
    #   uvicorn
requests==2.33.1
    # via
    #   langsmith
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
sniffio==1.3.1
    # via openai
soupsieve==2.8.3
    # via beautifulsoup4
starlette==1.0.0
    # via fastapi
tenacity==9.1.4
    # via langchain-core
tqdm==4.67.3
    # via openai
typing-extensions==4.15.0
    # via
    #   anyio
    #   beautifulsoup4
    #   fastapi
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
urllib3==2.6.3
    # via requests
uuid-utils==0.14.1
    # via
    #   langchain-core
    #   langsmith
uvicorn[standard]==0.42.0
    # via -r requirements.in
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
xxhash==3.6.0
    # via
    #   langgraph
    #   langsmith
zstandard==0.25.0
    # via langsmith
